created tick command processing colony production

+++ goods/models.py
@@ -26,17 +26,73 @@ class Stock(models.Model):
 		with transaction.atomic():
 			for ga in other.goodassignment_set.all():
 				self_ga = self.goodassignment_set.filter(good=ga.good)
+
 				if self_ga.count() < 1:
 					#this stock does not contain this good
-					raise Exception("could not deduct resource from this stock")
 
-				#deduct resource
-				self_ga = self_ga.first()
-				if self_ga.count == ga.count:
-					self_ga.delete()
+					if ga.count > 0:
+						#this would create a negative count
+						raise Exception("could not deduct resource from this stock")
+
+					#add new ressource
+					if ga.count == 0:
+						#nothing to do
+						continue
+
+					#create new goodassignment for this stock
+					self.goodassignment_set.create(good=ga.good, count=-ga.count)
+
 				else:
-					self_ga.count -= ga.count
-					self_ga.save()
+					#good is already in this stock
+
+					self_ga = self_ga.first()
+					if self_ga.count - ga.count < 0:
+						#this stock does not contain enough of this ressource
+						raise Exception("could not deduct resource from this stock")
+
+					#deduct resource
+					if self_ga.count == ga.count:
+						self_ga.delete()
+					else:
+						self_ga.count -= ga.count
+						self_ga.save()
+		return self
+
+	def __iadd__(self, other):
+		"""implements Stock += Stock"""
+		with transaction.atomic():
+			for ga in other.goodassignment_set.all():
+				self_ga = self.goodassignment_set.filter(good=ga.good)
+
+				if self_ga.count() < 1:
+					#this stock does not contain this good
+
+					if ga.count < 0:
+						#this would create a negative count
+						raise Exception("could not add new negative resource from this stock")
+
+					#add new ressource
+					if ga.count == 0:
+						#nothing to do
+						continue
+
+					#create new goodassignment for this stock
+					self.goodassignment_set.create(good=ga.good, count=ga.count)
+
+				else:
+					#good is already in this stock
+
+					self_ga = self_ga.first()
+					if self_ga.count + ga.count < 0:
+						#this stock does not contain enough of this ressource
+						raise Exception("could not add negative resource to this stock")
+
+					#deduct resource
+					if self_ga.count == ga.count:
+						self_ga.delete()
+					else:
+						self_ga.count += ga.count
+						self_ga.save()
 		return self
 
 	def __str__(self):
